gdbserver: recover software breakpoints on error
software breakpoints are implemented by replacing some instruction
directly in flash with the BREAK instruction. this means that we need to
restore the original instruction after reaching the breakpoint and/or
when we get some signal and need to exit, otherwise the flash would be
corrupted.

we try our best here to restore the instructions, but errors may happen,
and some signals can't be catched. please be careful, and use the
`verify` command to make sure that flash is consistent.